npm, nuget renames
