+++ lib/npm-provider.js
