+++ lib/nuget-provider.js
